images_api/requirements.txt
boto3
flask
gunicorn
Pillow
pytest
pytest-flask
pytest-mock
